feat: tidy up helper script
  Closes #1068
